First proof of concept, all loading and adjective fixing completed
While not a game yet, the majority of the generation is out of the way.
The next commit will contain code handling a basic combat system and
enemy generation based on the files. Later on will make a directory for
data files although C stdlib seems to struggle with this in a way python
didn't with the original